Adds docs, reverts structure for private
